feat: make use of suspense to improve initial rendering time
