+++ kit/dapp/src/app/[locale]/(private)/assets/(dashboard)/page.tsx
@@ -1,8 +1,10 @@
 import { PageHeader } from "@/components/layout/page-header";
+import { Skeleton } from "@/components/ui/skeleton";
 import { metadata } from "@/lib/config/metadata";
 import type { Metadata } from "next";
 import type { Locale } from "next-intl";
 import { getTranslations } from "next-intl/server";
+import { Suspense } from "react";
 import { AssetActivity } from "./_components/charts/asset-activity";
 import { AssetsSupply } from "./_components/charts/assets-supply";
 import { TransactionsHistory } from "./_components/charts/transactions-history";
@@ -50,17 +52,33 @@ export default async function AdminDashboard({
     <>
       <PageHeader title={t("title")} section={t("asset-management")} />
       <div className="grid grid-cols-1 gap-4 divide-x-0 divide-y lg:grid-cols-4 lg:divide-x lg:divide-y-0">
-        <AssetsWidget />
-        <TransactionsWidget />
-        <UsersWidget />
-        <PriceWidget />
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <AssetsWidget />
+        </Suspense>
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <TransactionsWidget />
+        </Suspense>
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <UsersWidget />
+        </Suspense>
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <PriceWidget />
+        </Suspense>
       </div>
       <p className="mt-8 mb-4 font-semibold text-2xl">{t("stats-heading")}</p>
       <div className="grid grid-cols-1 gap-4 divide-x-0 divide-y lg:grid-cols-2 lg:divide-x lg:divide-y-0 2xl:grid-cols-4">
-        <AssetsSupply />
-        <AssetActivity />
-        <UsersHistory />
-        <TransactionsHistory />
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <AssetsSupply />
+        </Suspense>
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <AssetActivity />
+        </Suspense>
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <UsersHistory />
+        </Suspense>
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <TransactionsHistory />
+        </Suspense>
       </div>
       <p className="mt-8 mb-4 font-semibold text-2xl">
         {t("latest-events-heading")}
