+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/(details)/_components/details/collateral.tsx
@@ -7,7 +7,6 @@ import { formatDate, formatDuration } from "@/lib/utils/date";
 import { formatNumber } from "@/lib/utils/number";
 import type { AssetType } from "@/lib/utils/typebox/asset-types";
 import { getLocale, getTranslations } from "next-intl/server";
-import { Suspense } from "react";
 import type { Address } from "viem";
 
 interface CollateralProps {
@@ -31,51 +30,49 @@ export async function Collateral({ address, assettype }: CollateralProps) {
   }
 
   return (
-    <Suspense>
-      <DetailGrid className="mt-4" title="Collateral">
-        <DetailGridItem
-          label={t("proven-collateral")}
-          info={t("proven-collateral-info")}
-        >
-          {formatNumber(asset.collateral, {
-            token: asset.symbol,
-            locale: locale,
-          })}
-        </DetailGridItem>
-        <DetailGridItem
-          label={t("required-collateral-threshold")}
-          info={t("required-collateral-threshold-info")}
-        >
-          {formatNumber(100, {
-            percentage: true,
-            decimals: 2,
-            locale: locale,
-          })}
-        </DetailGridItem>
-        <DetailGridItem
-          label={t("committed-collateral-ratio")}
-          info={t("committed-collateral-ratio-info")}
-        >
-          <PercentageProgressBar percentage={asset.collateralRatio} />
-        </DetailGridItem>
-        <DetailGridItem
-          label={t("collateral-proof-expiration")}
-          info={t("collateral-proof-expiration-info")}
-        >
-          {asset.collateralProofValidity
-            ? formatDate(asset.collateralProofValidity, {
-                type: "absolute",
-                locale: locale,
-              })
-            : "-"}
-        </DetailGridItem>
-        <DetailGridItem
-          label={t("collateral-proof-validity")}
-          info={t("collateral-proof-validity-info")}
-        >
-          {formatDuration(asset.liveness.toString())}
-        </DetailGridItem>
-      </DetailGrid>
-    </Suspense>
+    <DetailGrid className="mt-4" title="Collateral">
+      <DetailGridItem
+        label={t("proven-collateral")}
+        info={t("proven-collateral-info")}
+      >
+        {formatNumber(asset.collateral, {
+          token: asset.symbol,
+          locale: locale,
+        })}
+      </DetailGridItem>
+      <DetailGridItem
+        label={t("required-collateral-threshold")}
+        info={t("required-collateral-threshold-info")}
+      >
+        {formatNumber(100, {
+          percentage: true,
+          decimals: 2,
+          locale: locale,
+        })}
+      </DetailGridItem>
+      <DetailGridItem
+        label={t("committed-collateral-ratio")}
+        info={t("committed-collateral-ratio-info")}
+      >
+        <PercentageProgressBar percentage={asset.collateralRatio} />
+      </DetailGridItem>
+      <DetailGridItem
+        label={t("collateral-proof-expiration")}
+        info={t("collateral-proof-expiration-info")}
+      >
+        {asset.collateralProofValidity
+          ? formatDate(asset.collateralProofValidity, {
+              type: "absolute",
+              locale: locale,
+            })
+          : "-"}
+      </DetailGridItem>
+      <DetailGridItem
+        label={t("collateral-proof-validity")}
+        info={t("collateral-proof-validity-info")}
+      >
+        {formatDuration(asset.liveness.toString())}
+      </DetailGridItem>
+    </DetailGrid>
   );
 }
