+++ kit/dapp/src/app/[locale]/(private)/assets/[assettype]/[address]/(details)/_components/details/stablecoins.tsx
@@ -5,7 +5,6 @@ import { getAssetBalanceDetail } from "@/lib/queries/asset-balance/asset-balance
 import { getStableCoinDetail } from "@/lib/queries/stablecoin/stablecoin-detail";
 import { formatNumber } from "@/lib/utils/number";
 import { getLocale, getTranslations } from "next-intl/server";
-import { Suspense } from "react";
 import type { Address } from "viem";
 
 interface StablecoinsDetailsProps {
@@ -32,86 +31,81 @@ export async function StablecoinsDetails({
       : null;
 
   return (
-    <Suspense>
-      <DetailGrid>
-        <DetailGridItem label={t("name")}>{stableCoin.name}</DetailGridItem>
-        <DetailGridItem label={t("symbol")}>{stableCoin.symbol}</DetailGridItem>
-        {stableCoin.isin && (
-          <DetailGridItem label={t("isin")}>{stableCoin.isin}</DetailGridItem>
-        )}
-        <DetailGridItem label={t("contract-address")}>
-          <EvmAddress
-            address={stableCoin.id}
-            prettyNames={false}
-            hoverCard={false}
-            copyToClipboard={true}
-          />
-        </DetailGridItem>
-        <DetailGridItem label={t("creator")}>
-          <EvmAddress
-            address={stableCoin.creator.id}
-            hoverCard={false}
-            copyToClipboard={true}
-          />
-        </DetailGridItem>
-        <DetailGridItem label={t("decimals")}>
-          {stableCoin.decimals}
-        </DetailGridItem>
-        <DetailGridItem label={t("total-supply")} info={t("total-supply-info")}>
-          {formatNumber(stableCoin.totalSupply, {
+    <DetailGrid>
+      <DetailGridItem label={t("name")}>{stableCoin.name}</DetailGridItem>
+      <DetailGridItem label={t("symbol")}>{stableCoin.symbol}</DetailGridItem>
+      {stableCoin.isin && (
+        <DetailGridItem label={t("isin")}>{stableCoin.isin}</DetailGridItem>
+      )}
+      <DetailGridItem label={t("contract-address")}>
+        <EvmAddress
+          address={stableCoin.id}
+          prettyNames={false}
+          hoverCard={false}
+          copyToClipboard={true}
+        />
+      </DetailGridItem>
+      <DetailGridItem label={t("creator")}>
+        <EvmAddress
+          address={stableCoin.creator.id}
+          hoverCard={false}
+          copyToClipboard={true}
+        />
+      </DetailGridItem>
+      <DetailGridItem label={t("decimals")}>
+        {stableCoin.decimals}
+      </DetailGridItem>
+      <DetailGridItem label={t("total-supply")} info={t("total-supply-info")}>
+        {formatNumber(stableCoin.totalSupply, {
+          token: stableCoin.symbol,
+          locale: locale,
+        })}
+      </DetailGridItem>
+      {/* Show balance only when requested and available */}
+      {showBalance && balanceData && (
+        <DetailGridItem label={t("balance")}>
+          {formatNumber(balanceData.value, {
             token: stableCoin.symbol,
             locale: locale,
           })}
         </DetailGridItem>
-        {/* Show balance only when requested and available */}
-        {showBalance && balanceData && (
-          <DetailGridItem label={t("balance")}>
-            {formatNumber(balanceData.value, {
-              token: stableCoin.symbol,
-              locale: locale,
-            })}
-          </DetailGridItem>
-        )}
-        <DetailGridItem label={t("total-burned")} info={t("total-burned-info")}>
-          {formatNumber(stableCoin.totalBurned, {
-            token: stableCoin.symbol,
-            locale: locale,
-          })}
-        </DetailGridItem>
-        <DetailGridItem
-          label={t("total-holders")}
-          info={t("total-holders-info")}
-        >
-          {formatNumber(stableCoin.totalHolders, {
-            decimals: 0,
-            locale: locale,
-          })}
-        </DetailGridItem>
-        <DetailGridItem
-          label={t("ownership-concentration")}
-          info={t("ownership-concentration-info")}
-        >
-          {formatNumber(stableCoin.concentration, {
-            percentage: true,
-            decimals: 2,
-            locale: locale,
-          })}
-        </DetailGridItem>
-        <DetailGridItem label={t("price")}>
-          {formatNumber(stableCoin.price.amount, {
-            currency: stableCoin.price.currency,
-            decimals: 2,
-            locale: locale,
-          })}
-        </DetailGridItem>
-        <DetailGridItem label={t("total-value")}>
-          {formatNumber(stableCoin.price.amount * stableCoin.totalSupply, {
-            currency: stableCoin.price.currency,
-            decimals: 2,
-            locale: locale,
-          })}
-        </DetailGridItem>
-      </DetailGrid>
-    </Suspense>
+      )}
+      <DetailGridItem label={t("total-burned")} info={t("total-burned-info")}>
+        {formatNumber(stableCoin.totalBurned, {
+          token: stableCoin.symbol,
+          locale: locale,
+        })}
+      </DetailGridItem>
+      <DetailGridItem label={t("total-holders")} info={t("total-holders-info")}>
+        {formatNumber(stableCoin.totalHolders, {
+          decimals: 0,
+          locale: locale,
+        })}
+      </DetailGridItem>
+      <DetailGridItem
+        label={t("ownership-concentration")}
+        info={t("ownership-concentration-info")}
+      >
+        {formatNumber(stableCoin.concentration, {
+          percentage: true,
+          decimals: 2,
+          locale: locale,
+        })}
+      </DetailGridItem>
+      <DetailGridItem label={t("price")}>
+        {formatNumber(stableCoin.price.amount, {
+          currency: stableCoin.price.currency,
+          decimals: 2,
+          locale: locale,
+        })}
+      </DetailGridItem>
+      <DetailGridItem label={t("total-value")}>
+        {formatNumber(stableCoin.price.amount * stableCoin.totalSupply, {
+          currency: stableCoin.price.currency,
+          decimals: 2,
+          locale: locale,
+        })}
+      </DetailGridItem>
+    </DetailGrid>
   );
 }
