+++ kit/dapp/src/app/[locale]/(private)/portfolio/(dashboard)/page.tsx
@@ -1,23 +1,18 @@
 import { AssetDistribution } from "@/components/blocks/charts/assets/asset-distribution";
-import { PortfolioValue } from "@/components/blocks/charts/portfolio/portfolio-value";
 import MyAssetsTable from "@/components/blocks/my-assets-table/my-assets-table-mini";
-import { TransactionsHistory } from "@/components/blocks/transactions-table/transactions-history";
 import { PageHeader } from "@/components/layout/page-header";
+import { Skeleton } from "@/components/ui/skeleton";
 import { getUser } from "@/lib/auth/utils";
 import { metadata } from "@/lib/config/metadata";
-import { getUserAssetsBalance } from "@/lib/queries/asset-balance/asset-balance-user";
-import { getAssetsPricesInUserCurrency } from "@/lib/queries/asset-price/asset-price";
-import { getPortfolioStats } from "@/lib/queries/portfolio/portfolio-stats";
-import { getTransactionsTimeline } from "@/lib/queries/transactions/transactions-timeline";
-import { getCurrentUserDetail } from "@/lib/queries/user/user-detail";
-import { startOfDay, subMonths } from "date-fns";
 import type { Metadata } from "next";
 import type { Locale } from "next-intl";
 import { getTranslations } from "next-intl/server";
+import { Suspense } from "react";
 import type { Address } from "viem";
 import { LatestEvents } from "../../assets/(dashboard)/_components/table/latest-events";
 import { Greeting } from "./_components/greeting/greeting";
 import { MyAssetsHeader } from "./_components/header/my-assets-header";
+import { TransactionsChart } from "./_components/transactions-chart";
 
 export async function generateMetadata({
   params,
@@ -45,39 +40,13 @@ export default async function PortfolioDashboard({
   params: Promise<{ locale: Locale }>;
 }) {
   const { locale } = await params;
-  const t = await getTranslations({
-    locale,
-    namespace: "portfolio",
-  });
-
-  const user = await getUser();
-  const oneMonthAgo = startOfDay(subMonths(new Date(), 1));
-
-  const [myAssetsBalance, transactionsData, userDetails, portfolioStats] =
-    await Promise.all([
-      getUserAssetsBalance(user.wallet as Address),
-      getTransactionsTimeline({
-        timelineStartDate: oneMonthAgo,
-        granularity: "DAY",
-        from: user.wallet as Address,
-      }),
-      getCurrentUserDetail(),
-      getPortfolioStats({
-        address: user.wallet as Address,
-        days: 30,
-      }),
-    ]);
-
-  const assetPrices = await getAssetsPricesInUserCurrency(
-    portfolioStats.map(({ asset }) => asset.id)
-  );
-
-  const totalUserAssetsValue = myAssetsBalance.balances.reduce(
-    (acc, balance) =>
-      acc +
-      (assetPrices.get(balance.asset.id)?.amount ?? 0) * Number(balance.value),
-    0
-  );
+  const [t, user] = await Promise.all([
+    getTranslations({
+      locale,
+      namespace: "portfolio",
+    }),
+    getUser(),
+  ]);
 
   return (
     <>
@@ -87,37 +56,26 @@ export default async function PortfolioDashboard({
       />
       <div className="space-y-4">
         <Greeting />
-        <MyAssetsHeader
-          data={myAssetsBalance}
-          userDetails={userDetails}
-          totalValue={{
-            amount: totalUserAssetsValue,
-            currency: userDetails.currency,
-          }}
-        />
-        <PortfolioValue
-          portfolioStats={portfolioStats}
-          assetPriceMap={assetPrices}
-          locale={locale}
-        />
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <MyAssetsHeader locale={locale} walletAddress={user.wallet} />
+        </Suspense>
       </div>
 
       <PageHeader title={t("dashboard.my-assets")} className="mt-8" />
       <div className="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-3">
-        <AssetDistribution address={user.wallet as Address} />
-        <MyAssetsTable
-          wallet={user.wallet as Address}
-          title={t("dashboard.my-assets")}
-          variant="small"
-        />
-        <TransactionsHistory
-          data={transactionsData}
-          chartOptions={{
-            intervalType: "month",
-            intervalLength: 1,
-            granularity: "day",
-          }}
-        />
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <AssetDistribution address={user.wallet} />
+        </Suspense>
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <MyAssetsTable
+            wallet={user.wallet}
+            title={t("dashboard.my-assets")}
+            variant="small"
+          />
+        </Suspense>
+        <Suspense fallback={<Skeleton className="h-full w-full bg-muted/50" />}>
+          <TransactionsChart walletAddress={user.wallet} />
+        </Suspense>
       </div>
       <PageHeader title={t("dashboard.latest-events")} className="mt-8 mb-4" />
       <LatestEvents sender={user.wallet as Address} />
